+++ packages/build-clean/src/core/types.ts
@@ -3,7 +3,7 @@ import type { Options } from '../types';
 
 type FullFeatures = Record<Feature, boolean>;
 
-type RequiredOptions = 'debug' | 'verbose' | 'destructive' | 'features' | 'logger';
+type RequiredOptions = 'debug' | 'verbose' | 'destructive' | 'strict' | 'features' | 'logger';
 
 type MakeRequired<T, K extends keyof T> = Omit<T, K> & Required<Pick<T, K>>;